Fix compile error

+++ run.sh
@@ -1,53 +1,48 @@
 #!/bin/bash
 
+RED='\033[31m'
+GREEN='\033[32m'
+RESET='\033[0m'
+
 display_elapsed_time() {
   end_time=$(date +%s)
   elapsed_time=$((end_time - start_time))
   echo -ne "Elapsed time: ${elapsed_time} seconds\r"
 }
 
-
 start_time=$(date +%s)
 
-#if command -v cc >/dev/null 2>&1; then
-#  echo "cc is already installed."
-#else
-#  echo "cc is not installed. Installing it..."
+cd pypkg && ./run && cd ..
 
-#  if command -v apt >/dev/null 2>&1; then
-#    sudo apt update
-#    sudo apt install -y cc
-#  else
-#    echo "Package manager not found. Please install 'cc' manually."
-#  fi
-#fi
+if gcc -Wall -Wextra -Werror Script.c ../ft_*.c -o Run; then
+  chmod +x Run
 
-gcc -Wall -Wextra -Werror Script.c ../ft_*.c -o Run
-#cc check_puts.c ../ft_putstr_fd.c -o puts_run
+  display_elapsed_time &
+  output=$(./Run)
+  end_time=$(date +%s)
 
-chmod +x Run
+  kill %1
 
-display_elapsed_time &
+  elapsed_time=$((end_time - start_time))
+  echo "Total execution time: ${elapsed_time} seconds"
+  echo ""
 
-cd pypkg &&  ./run
-cd ..
+  echo "$output"
 
-output=$(./Run)
-end_time=$(date +%s)
+  if [ -e "./Run" ]; then
+    rm Run
+  fi
 
-kill %1
+  if gcc -Wall -Wextra -Werror ./fileDescriptor/check_puts.c ../ft_putstr_fd.c -o puts_run; then
+    ./puts_run | cat -e
+    echo ""
+    rm puts_run
+  else
+    echo -e "${RED}Compilation of puts_run failed. Skipping.${RESET}"
+  fi
+else
+  echo -e "${RED}Compilation of Run failed. Skipping.${RESET}"
+fi
 
-elapsed_time=$((end_time - start_time))
-echo "Total execution time: ${elapsed_time} seconds"
-echo ""
-
-echo "$output"
-
-cc ./fileDescriptor/check_puts.c ../ft_putstr_fd.c -o puts_run
-./puts_run | cat -e
-
-echo ""
-
-rm Run puts_run
-
-alias run=./Script.sh
+alias run=./run.sh
+echo -e "${GREEN}done."
